Entity-caching v2: adding support for timer/suspended/DL job count on execution.

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/persistence/entity/ExecutionEntityManagerImpl.java
@@ -540,12 +540,16 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
     }
 
     // Delete jobs
-    TimerJobEntityManager timerJobEntityManager = getTimerJobEntityManager();
-    Collection<TimerJobEntity> timerJobsForExecution = timerJobEntityManager.findJobsByExecutionId(executionEntity.getId());
-    for (TimerJobEntity job : timerJobsForExecution) {
-      timerJobEntityManager.delete(job);
-      if (getEventDispatcher().isEnabled()) {
-        getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+    
+    if (!enableExecutionRelationshipCounts 
+        || (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getTimerJobCount() > 0)) {
+      TimerJobEntityManager timerJobEntityManager = getTimerJobEntityManager();
+      Collection<TimerJobEntity> timerJobsForExecution = timerJobEntityManager.findJobsByExecutionId(executionEntity.getId());
+      for (TimerJobEntity job : timerJobsForExecution) {
+        timerJobEntityManager.delete(job);
+        if (getEventDispatcher().isEnabled()) {
+          getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+        }
       }
     }
     
@@ -561,21 +565,27 @@ public class ExecutionEntityManagerImpl extends AbstractEntityManager<ExecutionE
       }
     }
     
-    SuspendedJobEntityManager suspendedJobEntityManager = getSuspendedJobEntityManager();
-    Collection<SuspendedJobEntity> suspendedJobsForExecution = suspendedJobEntityManager.findJobsByExecutionId(executionEntity.getId());
-    for (SuspendedJobEntity job : suspendedJobsForExecution) {
-      suspendedJobEntityManager.delete(job);
-      if (getEventDispatcher().isEnabled()) {
-        getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+    if (!enableExecutionRelationshipCounts
+        || (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getSuspendedJobCount() > 0)) {
+      SuspendedJobEntityManager suspendedJobEntityManager = getSuspendedJobEntityManager();
+      Collection<SuspendedJobEntity> suspendedJobsForExecution = suspendedJobEntityManager.findJobsByExecutionId(executionEntity.getId());
+      for (SuspendedJobEntity job : suspendedJobsForExecution) {
+        suspendedJobEntityManager.delete(job);
+        if (getEventDispatcher().isEnabled()) {
+          getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+        }
       }
     }
     
-    DeadLetterJobEntityManager deadLetterJobEntityManager = getDeadLetterJobEntityManager();
-    Collection<DeadLetterJobEntity> deadLetterJobsForExecution = deadLetterJobEntityManager.findJobsByExecutionId(executionEntity.getId());
-    for (DeadLetterJobEntity job : deadLetterJobsForExecution) {
-      deadLetterJobEntityManager.delete(job);
-      if (getEventDispatcher().isEnabled()) {
-        getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+    if (!enableExecutionRelationshipCounts
+        || (enableExecutionRelationshipCounts && ((CountingExecutionEntity) executionEntity).getDeadLetterJobCount() > 0)) {
+      DeadLetterJobEntityManager deadLetterJobEntityManager = getDeadLetterJobEntityManager();
+      Collection<DeadLetterJobEntity> deadLetterJobsForExecution = deadLetterJobEntityManager.findJobsByExecutionId(executionEntity.getId());
+      for (DeadLetterJobEntity job : deadLetterJobsForExecution) {
+        deadLetterJobEntityManager.delete(job);
+        if (getEventDispatcher().isEnabled()) {
+          getEventDispatcher().dispatchEvent(ActivitiEventBuilder.createEntityEvent(ActivitiEventType.JOB_CANCELED, job));
+        }
       }
     }
